refactor(runtimed): extract runtimed-client crate from runtimed
Move client, protocol, service, sync_client, runtime, settings_doc,
and singleton modules into a new runtimed-client crate. This breaks
the dependency of notebook (Tauri app) on runtimed internals, allowing
the app crate to use the lightweight client without pulling in daemon
code or rattler dependencies. Also makes kernel-env progress types
available without feature gating.
