+++ crates/notebook-sync/src/tests.rs
@@ -602,7 +602,7 @@ mod integration_tests {
         if let Ok(path) = std::env::var("RUNTIMED_SOCKET_PATH") {
             return PathBuf::from(path);
         }
-        runtimed::default_socket_path()
+        runtimed_client::default_socket_path()
     }
 
     /// Check if a daemon is running and available.
