+++ crates/runt/Cargo.toml
@@ -21,8 +21,8 @@ serde_json = { workspace = true }
 uuid = { workspace = true }
 jupyter-protocol = { workspace = true }
 runtimelib = { workspace = true, features = ["tokio-runtime", "aws-lc-rs"] }
-runtimed = { path = "../runtimed" }
-notebook-doc = { path = "../notebook-doc" }
+runtimed-client = { path = "../runtimed-client" }
+notebook-doc = { path = "../notebook-doc", features = ["persistence"] }
 runt-workspace = { path = "../runt-workspace" }
 kernel-env = { path = "../kernel-env" }
 clap = { version = "4.5.1", features = ["derive", "color"] }
