+++ crates/runtimed/src/singleton.rs
@@ -1,35 +1,20 @@
 //! Singleton management for the pool daemon.
 //!
 //! Ensures only one daemon instance runs per user using file-based locking.
+//! Read-only daemon discovery (DaemonInfo, get_running_daemon_info, etc.)
+//! lives in `runtimed_client::singleton` and is re-exported via `pub use runtimed_client::*`.
 
 use std::fs::{File, OpenOptions};
 use std::path::PathBuf;
 
-use chrono::{DateTime, Utc};
+use chrono::Utc;
 use log::{info, warn};
-use serde::{Deserialize, Serialize};
 
-/// Information about a running daemon instance.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DaemonInfo {
-    /// Socket endpoint the daemon is listening on.
-    pub endpoint: String,
-    /// Process ID of the daemon.
-    pub pid: u32,
-    /// Version of the daemon.
-    pub version: String,
-    /// When the daemon started.
-    pub started_at: DateTime<Utc>,
-    /// HTTP port for the blob server (if running).
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub blob_port: Option<u16>,
-    /// Path to the git worktree (dev mode only).
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub worktree_path: Option<String>,
-    /// Human-readable workspace description (dev mode only).
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub workspace_description: Option<String>,
-}
+// Re-export all client-side singleton items so `runtimed::singleton::*` still works.
+use runtimed_client::singleton as client_singleton;
+pub use runtimed_client::singleton::{
+    daemon_info_path, daemon_lock_path, get_running_daemon_info, read_daemon_info, DaemonInfo,
+};
 
 /// A lock that ensures only one daemon instance runs.
 pub struct DaemonLock {
@@ -46,11 +31,16 @@ impl DaemonLock {
     ///
     /// If `custom_lock_dir` is provided, uses that directory for lock files
     /// instead of the default. This is primarily for testing.
-    pub fn try_acquire(custom_lock_dir: Option<&PathBuf>) -> Result<Self, DaemonInfo> {
+    pub fn try_acquire(
+        custom_lock_dir: Option<&PathBuf>,
+    ) -> Result<Self, client_singleton::DaemonInfo> {
         let (lock_path, info_path) = if let Some(dir) = custom_lock_dir {
             (dir.join("daemon.lock"), dir.join("daemon.json"))
         } else {
-            (daemon_lock_path(), daemon_info_path())
+            (
+                client_singleton::daemon_lock_path(),
+                client_singleton::daemon_info_path(),
+            )
         };
 
         // Ensure parent directory exists
@@ -69,11 +59,11 @@ impl DaemonLock {
             Err(e) => {
                 warn!("[singleton] Failed to open lock file: {}", e);
                 // Try to read existing daemon info
-                if let Some(info) = read_daemon_info(&info_path) {
+                if let Some(info) = client_singleton::read_daemon_info(&info_path) {
                     return Err(info);
                 }
                 // No info available, create a placeholder
-                return Err(DaemonInfo {
+                return Err(client_singleton::DaemonInfo {
                     endpoint: "unknown".to_string(),
                     pid: 0,
                     version: "unknown".to_string(),
@@ -94,10 +84,10 @@ impl DaemonLock {
             if result != 0 {
                 // Another process holds the lock
                 info!("[singleton] Another daemon is already running");
-                if let Some(info) = read_daemon_info(&info_path) {
+                if let Some(info) = client_singleton::read_daemon_info(&info_path) {
                     return Err(info);
                 }
-                return Err(DaemonInfo {
+                return Err(client_singleton::DaemonInfo {
                     endpoint: "unknown".to_string(),
                     pid: 0,
                     version: "unknown".to_string(),
@@ -132,10 +122,10 @@ impl DaemonLock {
             };
             if result == 0 {
                 info!("[singleton] Another daemon is already running");
-                if let Some(info) = read_daemon_info(&info_path) {
+                if let Some(info) = client_singleton::read_daemon_info(&info_path) {
                     return Err(info);
                 }
-                return Err(DaemonInfo {
+                return Err(client_singleton::DaemonInfo {
                     endpoint: "unknown".to_string(),
                     pid: 0,
                     version: "unknown".to_string(),
@@ -168,7 +158,7 @@ impl DaemonLock {
             (None, None)
         };
 
-        let info = DaemonInfo {
+        let info = client_singleton::DaemonInfo {
             endpoint: endpoint.to_string(),
             pid: std::process::id(),
             version: crate::daemon_version().to_string(),
@@ -199,43 +189,5 @@ impl Drop for DaemonLock {
             std::fs::remove_file(&self.info_path).ok();
         }
         info!("[singleton] Released daemon lock");
-    }
-}
-
-/// Get the path to the daemon lock file.
-pub fn daemon_lock_path() -> PathBuf {
-    crate::daemon_base_dir().join("daemon.lock")
-}
-
-/// Get the path to the daemon info file.
-pub fn daemon_info_path() -> PathBuf {
-    crate::daemon_base_dir().join("daemon.json")
-}
-
-/// Read daemon info from the info file.
-pub fn read_daemon_info(path: &PathBuf) -> Option<DaemonInfo> {
-    std::fs::read_to_string(path)
-        .ok()
-        .and_then(|s| serde_json::from_str(&s).ok())
-}
-
-/// Check if a daemon is running by reading the info file.
-pub fn get_running_daemon_info() -> Option<DaemonInfo> {
-    read_daemon_info(&daemon_info_path())
-}
-
-#[cfg(test)]
-#[allow(clippy::unwrap_used, clippy::expect_used)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_daemon_paths() {
-        let lock_path = daemon_lock_path();
-        let info_path = daemon_info_path();
-
-        assert!(lock_path.to_string_lossy().contains("runt"));
-        assert!(lock_path.to_string_lossy().contains("daemon.lock"));
-        assert!(info_path.to_string_lossy().contains("daemon.json"));
     }
 }
